Refactor: general reworks

+++ lib/tests/controller.test.ts
@@ -1,5 +1,5 @@
 import { newUser, authenticateUser } from '../controllers/users';
-import { User } from '../db';
+import { User, DBUserInterface } from '../db';
 import mongoose from 'mongoose';
 
 const status200 = 200;
@@ -9,7 +9,18 @@ const status404 = 404;
 
 describe('controller', () => {
     test('create new user', async done => {
-        const user = await newUser({
+        const user:
+            | {
+                  status: number;
+                  error: string;
+                  message?: undefined;
+              }
+            | {
+                  status: number;
+                  message: string;
+                  error?: undefined;
+              }
+            | undefined = await newUser({
             email: 'test@testing.com',
             firstName: 'John',
             lastName: 'Doe',
@@ -25,7 +36,18 @@ describe('controller', () => {
     });
 
     test('create an existing user', async done => {
-        const user = await newUser({
+        const user:
+            | {
+                  status: number;
+                  error: string;
+                  message?: undefined;
+              }
+            | {
+                  status: number;
+                  message: string;
+                  error?: undefined;
+              }
+            | undefined = await newUser({
             email: 'test@testing.com',
             firstName: 'John',
             lastName: 'Doe',
@@ -38,28 +60,80 @@ describe('controller', () => {
     });
 
     test('authenticate existing user', async done => {
-        const auth: any = await authenticateUser({
+        const auth:
+            | {
+                  status: number;
+                  error: string;
+                  message?: undefined;
+                  user?: undefined;
+                  token?: undefined;
+              }
+            | {
+                  status: number;
+                  message: string;
+                  user: {
+                      username: string;
+                      email: string;
+                      role: string;
+                      lastName: string;
+                      firstName: string;
+                      createdAt: Date;
+                      updatedAt?: Date | undefined;
+                  };
+                  token: string;
+                  error?: undefined;
+              }
+            | undefined = await authenticateUser({
             password: 'testing123',
             username: 'jd12345',
         });
-        expect(auth.status).toEqual(status200);
+        if (auth) {
+            expect(auth.status).toEqual(status200);
+        }
         done();
     });
 
     test('authenticate a non existing user', async done => {
-        const auth: any = await authenticateUser({
+        const auth:
+            | {
+                  status: number;
+                  error: string;
+                  message?: undefined;
+                  user?: undefined;
+                  token?: undefined;
+              }
+            | {
+                  status: number;
+                  message: string;
+                  user: {
+                      username: string;
+                      email: string;
+                      role: string;
+                      lastName: string;
+                      firstName: string;
+                      createdAt: Date;
+                      updatedAt?: Date | undefined;
+                  };
+                  token: string;
+                  error?: undefined;
+              }
+            | undefined = await authenticateUser({
             password: 'testing123',
             username: 'dj12345',
         });
-        expect(auth.status).toEqual(status404);
+        if (auth) {
+            expect(auth.status).toEqual(status404);
+        }
         done();
     });
 
     afterAll(async done => {
-        const user: any = await User.findOne({
+        const user: DBUserInterface | null = await User.findOne({
             email: 'test@testing.com',
         });
-        await User.findByIdAndDelete(user._id);
+        if (user) {
+            await User.findByIdAndDelete(user._id);
+        }
         await mongoose.connection.close();
         done();
     });
